+++ company-gtags.el
@@ -48,6 +48,7 @@
   :package-version '(company . "0.8.1"))
 
 (defvar-local company-gtags--tags-available-p 'unknown)
+(defvar-local company-gtags--executable 'unknown)
 
 (defcustom company-gtags-modes '(prog-mode jde-mode)
   "Modes that use `company-gtags'.
@@ -61,6 +62,32 @@ completion."
       (setq company-gtags--tags-available-p
             (locate-dominating-file buffer-file-name "GTAGS"))
     company-gtags--tags-available-p))
+
+(defun company-gtags--executable ()
+  (cond
+   ((not (eq company-gtags--executable 'unknown)) ;; the value is already cached
+    company-gtags--executable)
+   ((and (version<= "27" emacs-version)           ;; can search remotely to set
+         (file-remote-p default-directory))
+
+    (with-connection-local-variables
+     (if (boundp 'company-gtags--executable-connection)
+         (setq-local company-gtags--executable     ;; use if defined as connection-local
+                     company-gtags--executable-connection)
+
+       ;; Else search and set as connection local for next uses.
+       (setq-local company-gtags--executable (executable-find "global" t))
+       (let* ((host (file-remote-p default-directory 'host))
+              (symvars (intern (concat host "-vars")))) ;; profile name
+
+         (connection-local-set-profile-variables
+          symvars
+          `((company-gtags--executable-connection . ,company-gtags--executable)))
+
+         (connection-local-set-profiles `(:machine ,host) symvars))
+       company-gtags--executable)))
+   (t     ;; use default value (searched locally)
+    company-gtags-executable)))
 
 (defun company-gtags--fetch-tags (prefix)
   (with-temp-buffer
@@ -98,7 +125,7 @@ completion."
   (interactive (list 'interactive))
   (cl-case command
     (interactive (company-begin-backend 'company-gtags))
-    (prefix (and company-gtags-executable
+    (prefix (and (company-gtags--executable)
                  buffer-file-name
                  (apply #'derived-mode-p company-gtags-modes)
                  (not (company-in-string-or-comment))